ai-service/requirements.txt
fastapi>=0.95.1
uvicorn>=0.22.0
python-dotenv>=1.0.0
pydantic>=1.10.7
transformers>=4.21.0,<5.0.0
torch>=1.12.0,<2.0.0
numpy>=1.21.6
pandas>=1.3.5
scikit-learn>=1.1.0